Started refactoring code.
Began with the bease styles and moving my way down. The style.css file was changed to a partial an imported to the main scss file. I will have to delete the styless.scss partial when I am done, but it allows me to make sure that nothing breaks as I am refactoring.